Update MSFT_xWebsite.psm1
The current xWebsite did not seem to accommodate hostheaders or multiple SSL on one site via hostheaders. The code is just adding a UseHostHeader flag that changes the Get-WebBinding to check for hostheader when binding to hostheadered SSL. This does not check if the SSLFlags setting is set correctly. 
This is my first time to contribute to GitHub, or opensource in general for that matter. So, hopefully I am doing this right.