Separated output from under-test and mocked cmdlet
Within each relevant test for Get-TargetResource, output from cmdlets that are used
as a part of Get-TargetResource is mocked (Get-Website and
Get-WebConfiguration). Originally the tests were comparing the actual
output of Get-TargetResource to the mocked output from these helper
cmdlets which has not yet been processed by Get-TargetResource. Now for
the LogFlags, the tests compare against the expected value of the
transformed (split and cast) output.