(#221) Inconsistent Logic in xWebsite
The Test-TargetResource function in the xWebSite resource used
inconsistent logic in the way that it would test a property for
compliance and then report non-compliance.

Some tests would use the `$PSBoundParameters` variable to check for
parameters to be tested, others would use `[String]::isNullOrEmpty()`

Some tests would set the `$inDesiredState` variable, write a verbose
message and move on, and others would write the verbose message and
use the `return` keyward to immediately exit the function and stop
testing additional properties.

The desired style should use `$PSBoundParameters` and the
`$inDesiredState` variable consistently, so that all properties are
tested for troubleshooting purposes, and to promote conciseness and
reasability

Resolves #221