DSC_WebApplication: Ensure class names match
Prior to this commit there as a mismatch between the class name for
WebApplicationAuthenticationInformation in the schema and
implementation.

The schema referenced the class as `DSC_WebApplicationAuthenticationInformation`
whereas the implementation referenced the class as
`MSDT_xWebApplicationAuthenticationInformation`.

This caused an error when CIM would later try and lookup the class. The
underlying error returned was `MI RESULT 5`.

When consumed by Puppet (via Puppet.Dsc) we see the following:

```
Error: dsc_webapplication: The specified Class does not exist.
Error: /Stage[main]/Main/Dsc_webapplication[TestApp]: Could not evaluate: undefined method `[]' for nil:NilClass
```

This commit resolves the issue by updating the class name passed to
`New-CimInstance` so that it matches the value defined in the schema.